requirements.txt
gunicorn
django-heroku
Pillow
social-auth-app-django
django-mptt
django-star-ratings
boto3
django-storages
django-tempus-dominus
django-crispy-forms
